requirements.txt
pandas==2.2.2
numpy==2.0.1
scipy==1.14.0
matplotlib==3.9.1
python-dotenv==1.0.1